Make sure that time stamps use 64 bits.

+++ src/components/ogre/Avatar.cpp
@@ -222,7 +222,7 @@ void Avatar::attemptMove()
 		S_LOG_VERBOSE(ss.str());
 
 		//Save the ten latest orientations sent to the server, so we can later when we receive an update from the server we can recognize that it's our own updates and ignore them.
-		long currentTime = Time::currentTimeMillis();
+		long long currentTime = Time::currentTimeMillis();
 		mLastTransmittedMovements.push_back(TimedMovementStateList::value_type(currentTime, newMovementState));
 		if (mLastTransmittedMovements.size() > 10) {
 			mLastTransmittedMovements.erase(mLastTransmittedMovements.begin());
@@ -249,7 +249,7 @@ bool Avatar::isOkayToSendRotationMovementChangeToServer()
 	if (!mLastTransmittedMovements.size()) {
 		return true;
 	}
-	long currentTime = Time::currentTimeMillis();
+	long long currentTime = Time::currentTimeMillis();
 	if ((currentTime - mLastTransmittedMovements.rbegin()->first) > mMinIntervalOfRotationChanges) {
 		return true;
 	}
@@ -369,7 +369,7 @@ WFMath::Point<3> Avatar::getClientSideAvatarPosition() const
 //	//	if (mCurrentMovement == WFMath::Vector<3>::ZERO() && mErisAvatarEntity.isMoving()) {
 //	//		bool clientSideMovement = false;
 //	//		if (mLastTransmittedMovements.size()) {
-//	//			long currentTime = EmberServices::getSingleton().getTimeService()->currentTimeMillis();
+//	//			long long currentTime = EmberServices::getSingleton().getTimeService()->currentTimeMillis();
 //	//			if ((currentTime - mLastTransmittedMovements.rbegin()->first) < 1000) {
 //	//				clientSideMovement = true;
 //	//			}
