+++ src/components/ogre/model/ModelDefinitionManager.h
@@ -123,7 +123,7 @@ public:
 	 * All of the background loaders will have their poll() method called. If the background loader has finished loading it will be removed from the store of loaders and the model will be reloaded.
 	 * @param maxTimeMilliseconds The max time allowed, presented as a future date, in unix time milliseconds.
 	 */
-	void pollBackgroundLoaders(long maxTimeMilliseconds);
+	void pollBackgroundLoaders(long long maxTimeMilliseconds);
 	
 protected:
