+++ src/framework/tasks/TaskQueue.h
@@ -79,7 +79,7 @@ public:
 	 * By setting maxAllowedTimeMilliseconds you can limit the amount of time the queue will spend on processed tasks. This is useful for keeping framerate up.
 	 * @param maxAllowedTimeMilliseconds If set to something larger than 0, this is the maximum milliseconds that the queue is allowed to process tasks.
 	 */
-	void pollProcessedTasks(long maxAllowedTimeMilliseconds);
+	void pollProcessedTasks(long long maxAllowedTimeMilliseconds);
 
 protected:
